UX improvements: collapsed defaults, short dates, double-click editing
- Default vendors and call notes sections to collapsed on project detail page
- Format dates as short month (e.g. "Mar 8, 2026") instead of long month
- Make auto-generated summaries editable by clicking on them inline
- Add EditableSummary component for inline summary editing in call notes
- Replace edit button with double-click to edit in NoteDetailModal (notes + summary)
- Add double-click on call notes from main page ClientCard to open editable modal
- Show "Double-click to edit" hint instead of edit button

+++ src/data/seed.ts
@@ -377,5 +377,5 @@ export function formatCountdown(eventDate: string): { text: string; isUrgent: bo
 
 export function formatDate(dateStr: string): string {
   const date = new Date(dateStr + 'T00:00:00');
-  return date.toLocaleDateString('en-US', { month: 'long', day: 'numeric', year: 'numeric' });
+  return date.toLocaleDateString('en-US', { month: 'short', day: 'numeric', year: 'numeric' });
 }
